:rocket: add divide the search in small components
