+++ src/app/pages/HomePage.jsx
@@ -4,7 +4,7 @@ import { usePlacesStore } from '../hooks/usePlacesStore';
 import { useActionsPlaces } from '../hooks/useActionsPlaces';
 import { ButtonLocationUser } from '../components/map/ButtonLocationUser';
 import Feedback from '../components/Feedback';
-import { Search } from '../components/Search';
+import { Header } from '../components/Header';
 
 const Homepage = () => {
   const { deniedLocation } = usePlacesStore();
@@ -12,7 +12,7 @@ const Homepage = () => {
 
   return (
     <>
-      <Search />
+      <Header />
       <MapView />
       {deniedLocation && <ErrorAccessLocation />}
       <ButtonLocationUser />
